Mayor recipe refactor
* Split functionalities into new recipes
* Use a simpler custom rvm recipe instead of external a cookbook
* Separate the databases into optional recipes
* Create a optional recipe for testing stuff (firefox, xvnc DISPLAY=:0, etc)
* Remove old external dependencies

+++ recipes/rvm.rb
@@ -1,35 +1,38 @@
-# Some handy vars
-user_home = "/home/#{node['bootstrap']['user']}"
-rvm_src_path = "#{user_home}/.rvm/src"
-user_env = {
-  'USER' => node['bootstrap']['user'],
-  'HOME' => user_home
-}
-
-
 node['bootstrap']['rvm']['packages'].each do |pkg|
   package pkg
 end
 
 execute "Install rvm" do
   command "curl -L #{node['bootstrap']['rvm']['installer_url']} | bash -s #{node['bootstrap']['rvm']['branch']}"
-  user node['bootstrap']['user']
-  cwd  user_home
-  environment user_env
-  creates "#{user_home}/.rvm/scripts/rvm"
+  user node['user_name']
+  cwd  node['user_home']
+  environment node['user_env']
+  creates "#{node['user_home']}/.rvm/scripts/rvm"
 end
 
 bash "Install ruby #{node['bootstrap']['rvm']['ruby']}" do
   code "source $HOME/.rvm/scripts/rvm ; rvm install #{node['bootstrap']['rvm']['ruby']}"
-  user node['bootstrap']['user']
-  cwd  user_home
-  environment user_env
-  creates "#{user_home}/.rvm/rubies/#{node['rvm']['ruby']}/config"
+  user node['user_name']
+  cwd  node['user_home']
+  environment node['user_env']
+  creates "#{node['user_home']}/.rvm/rubies/#{node['bootstrap']['rvm']['ruby']}/config"
   subscribes :run, resources(:execute => "Install rvm")
 end
 
-execute "remove rvm sources" do
-  command "rm -fr #{rvm_src_path}/*"
-  only_if { ::File.exists? "#{rvm_src_path}/rvm" }
-  subscribes :run, resources(:bash => "Install ruby #{node['bootstrap']['rvm']['ruby']}")
+if node['bootstrap']['rvm']['clean_sources']
+  rvm_src_path = "#{node['user_home']}/.rvm/src"
+  execute "remove rvm sources" do
+    command "rm -fr #{rvm_src_path}/*"
+    only_if { ::File.exists? "#{rvm_src_path}/rvm" }
+    subscribes :run, resources(:bash => "Install ruby #{node['bootstrap']['rvm']['ruby']}")
+  end
+end
+
+# Create a .gemrc for the aentos user
+file "#{node['user_home']}/.gemrc" do
+  content node['bootstrap']['gem_options']
+  owner  node['user_name']
+  group  node['user_name']
+  mode   0644
+  action :create_if_missing
 end
